% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{build_sparse_kernel}
\alias{build_sparse_kernel}
\title{Build sparse kNN kernel (exponential)
Build sparse k-NN Gibbs kernel}
\usage{
build_sparse_kernel(coords, k = 60, eps = 50)
}
\arguments{
\item{coords}{numeric matrix or data frame (eg. \code{lon}, \code{lat} columns)}

\item{k}{integer; number of neighbours}

\item{eps}{numeric; distance scale}
}
\value{
lista: \code{K} (dgCMatrix, \eqn{n \times n}) és \code{D} (distance matrix)
}
\description{
Build sparse kNN kernel (exponential)
Build sparse k-NN Gibbs kernel
}
