% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craneOT-package.R
\docType{package}
\name{craneOT-package}
\alias{craneOT}
\alias{craneOT-package}
\title{craneOT: Optimal Transport on k-NN Graphs for Roost Counts}
\description{
Tools for calculating the rearrangement from weekly roost surveys
(unbalanced, entropic regulrised OT)'
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vegvarizs/craneOT}
  \item Report bugs at \url{https://github.com/vegvarizs/craneOT/issues}
}

}
\author{
\strong{Maintainer}: Zsolt Végvári \email{zsolt.vegvari@gmail.com} (\href{https://orcid.org/0000-0002-2804-9282}{ORCID})

}
\keyword{internal}
