% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{build_sparse_kernel}
\alias{build_sparse_kernel}
\title{Build sparse kNN kernel (exponential)}
\usage{
build_sparse_kernel(coords, k = 60, eps = 50)
}
\arguments{
\item{coords}{matrix/data.frame with columns lon,lat (WGS84 or planar)}

\item{k}{integer, neighbors}

\item{eps}{numeric, distance scale (same units as coords distances)}
}
\value{
list(K=dgCMatrix, D=matrix)
}
\description{
Build sparse kNN kernel (exponential)
}
