% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows_metrics.R
\name{flows_summary}
\alias{flows_summary}
\title{Aggregate flows and compute hub scores & mean distance}
\usage{
flows_summary(flows, Dmat = NULL, A = NULL)
}
\arguments{
\item{flows}{list of n x n (dgCMatrix) weekly flows}

\item{Dmat}{optional n x n distances for mean distance}

\item{A}{optional n x T counts to scale hub score}
}
\description{
Aggregate flows and compute hub scores & mean distance
}
