% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinkhorn_uot.R
\name{sinkhorn_uot}
\alias{sinkhorn_uot}
\title{Sparse unbalanced Sinkhorn OT}
\usage{
sinkhorn_uot(a, b, K, eps = 60, rho = 30, maxiter = 1500, tol = 1e-06)
}
\arguments{
\item{a, b}{numeric vectors (length n)}

\item{K}{sparse kernel (dgCMatrix, n x n)}

\item{eps, rho}{positive scalars}
}
\value{
sparse transport plan (dgCMatrix)
}
\description{
Sparse unbalanced Sinkhorn OT
}
